chore: move User model to Domain and add DDD skeleton
