+++ app/Application/README.md
@@ -1,0 +1,9 @@
+Application 層
+
+このディレクトリはユースケースや DTO を配置します。
+
+例:
+- UseCases/
+- DTOs/
+
+（初期は雛形のみ）
