+++ app/Domain/Services/README.md
@@ -1,0 +1,5 @@
+Domain Services
+
+ドメインサービスを配置します。振る舞いは Entity にできる限り持たせ、残りをここに実装します。
+
+（初期は雛形のみ）
