+++ app/Domain/ValueObjects/README.md
@@ -1,0 +1,5 @@
+ValueObjects
+
+このディレクトリは `Money`, `Period`, `ProjectStatus` などの不変の値オブジェクトを配置します。
+
+（初期は雛形のみ）
