+++ app/Infrastructure/README.md
@@ -1,0 +1,8 @@
+Infrastructure 層
+
+このディレクトリはリポジトリ実装や外部連携のコードを配置します。
+
+例:
+- Repositories/
+
+（初期は雛形のみ）
